Move test to DefaultLongTaskTimerTest
It's testing logic in the DefaultLongTaskTimer class, so it is more appropriate for it to be in that test class.
