+++ micrometer-core/src/test/java/io/micrometer/core/instrument/internal/CumulativeHistogramLongTaskTimerTest.java
@@ -19,16 +19,9 @@ import io.micrometer.core.instrument.LongTaskTimer;
 import io.micrometer.core.instrument.Meter;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.MockClock;
-import io.micrometer.core.instrument.distribution.CountAtBucket;
 import io.micrometer.core.instrument.distribution.DistributionStatisticConfig;
 import io.micrometer.core.instrument.simple.SimpleConfig;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
-import org.junit.jupiter.api.Test;
-
-import java.time.Duration;
-import java.util.concurrent.TimeUnit;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Tests for {@link CumulativeHistogramLongTaskTimer}.
@@ -42,25 +35,4 @@ class CumulativeHistogramLongTaskTimerTest {
             return new CumulativeHistogramLongTaskTimer(id, clock, getBaseTimeUnit(), distributionStatisticConfig);
         }
     };
-
-    @Test
-    void histogramWithMoreBucketsThanActiveTasks() {
-        LongTaskTimer ltt = LongTaskTimer.builder("my.ltt").publishPercentileHistogram().register(registry);
-        ltt.start();
-        clock.add(15, TimeUnit.MINUTES);
-        ltt.start();
-        clock.add(5, TimeUnit.MINUTES);
-        // one task at 20 minutes, one task at 5 minutes
-        CountAtBucket[] countAtBuckets = ltt.takeSnapshot().histogramCounts();
-        int index = 0;
-        while (countAtBuckets[index].bucket(TimeUnit.NANOSECONDS) < Duration.ofMinutes(5).toNanos()) {
-            assertThat(countAtBuckets[index++].count()).isZero();
-        }
-        while (countAtBuckets[index].bucket(TimeUnit.NANOSECONDS) < Duration.ofMinutes(20).toNanos()) {
-            assertThat(countAtBuckets[index++].count()).isOne();
-        }
-        while (index < countAtBuckets.length) {
-            assertThat(countAtBuckets[index++].count()).isEqualTo(2);
-        }
-    }
 }
