+++ micrometer-test/src/main/java/io/micrometer/core/tck/DefaultLongTaskTimerTest.java
@@ -16,9 +16,12 @@
 package io.micrometer.core.tck;
 
 import io.micrometer.core.instrument.LongTaskTimer;
+import io.micrometer.core.instrument.Meter;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.MockClock;
 import io.micrometer.core.instrument.distribution.CountAtBucket;
+import io.micrometer.core.instrument.distribution.DistributionStatisticConfig;
+import io.micrometer.core.instrument.internal.DefaultLongTaskTimer;
 import io.micrometer.core.instrument.simple.SimpleConfig;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import org.junit.jupiter.api.DisplayName;
@@ -34,11 +37,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 public class DefaultLongTaskTimerTest {
 
-    MeterRegistry registry = new SimpleMeterRegistry(SimpleConfig.DEFAULT, new MockClock());
-
     @Test
     @DisplayName("supports sending histograms of active task duration")
     void histogram() {
+        MeterRegistry registry = new SimpleMeterRegistry(SimpleConfig.DEFAULT, new MockClock());
         LongTaskTimer t = LongTaskTimer.builder("my.timer")
                 .serviceLevelObjectives(Duration.ofSeconds(10), Duration.ofSeconds(40), Duration.ofMinutes(1))
                 .register(registry);
@@ -62,5 +64,34 @@ public class DefaultLongTaskTimerTest {
 
         assertThat(countAtBuckets[2].bucket(TimeUnit.MINUTES)).isEqualTo(1);
         assertThat(countAtBuckets[2].count()).isEqualTo(11);
+    }
+
+    @Test
+    void histogramWithMoreBucketsThanActiveTasks() {
+        MockClock clock = new MockClock();
+        MeterRegistry registry = new SimpleMeterRegistry(SimpleConfig.DEFAULT, clock) {
+            @Override
+            protected LongTaskTimer newLongTaskTimer(Meter.Id id, DistributionStatisticConfig distributionStatisticConfig) {
+                // supportsAggregablePercentiles true for using pre-defined histogram buckets
+                return new DefaultLongTaskTimer(id, clock, getBaseTimeUnit(), distributionStatisticConfig, true);
+            }
+        };
+        LongTaskTimer ltt = LongTaskTimer.builder("my.ltt").publishPercentileHistogram().register(registry);
+        ltt.start();
+        clock.add(15, TimeUnit.MINUTES);
+        ltt.start();
+        clock.add(5, TimeUnit.MINUTES);
+        // one task at 20 minutes, one task at 5 minutes
+        CountAtBucket[] countAtBuckets = ltt.takeSnapshot().histogramCounts();
+        int index = 0;
+        while (countAtBuckets[index].bucket(TimeUnit.NANOSECONDS) < Duration.ofMinutes(5).toNanos()) {
+            assertThat(countAtBuckets[index++].count()).isZero();
+        }
+        while (countAtBuckets[index].bucket(TimeUnit.NANOSECONDS) < Duration.ofMinutes(20).toNanos()) {
+            assertThat(countAtBuckets[index++].count()).isOne();
+        }
+        while (index < countAtBuckets.length) {
+            assertThat(countAtBuckets[index++].count()).isEqualTo(2);
+        }
     }
 }
